feat: add batching to figma export
