+++ scripts/figma_icons/types.ts
@@ -1,5 +1,11 @@
 export interface FigmaIconConfig {
   /**
+   * The number of images to process at a time
+   *
+   */
+  batchSize: number;
+
+  /**
    * Branch that is located in the
    * Figma file
    * @example spike-branch
